Move expiry tests to a separate module
